requirements.txt
brotli==1.0.9
fastapi==0.115.6
imblearn==0.0
importlib-metadata==8.0.0
importlib-resources==6.4.0
jaraco.collections==5.1.0
notebook==7.2.2
opencv-python==4.11.0.86
pip-chill==1.0.3
pipreqs==0.5.0
pysocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
seaborn==0.13.2
sklearn-compat==0.1.3
tensorflow==2.18.0
tree-cli==0.1.1
uvicorn==0.34.0
xgboost==2.1.0
